+++ CHANGELOG.md
@@ -6,12 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-
 ### Added
+
 - **CUMULUS-2591**
   - Adds `failedExecutionStepName` to failed execution's jsonb error records.
     This is the name of the Step Function step for the last failed event in the
     execution's event history.
+
+### Fixed
+
+- **CUMULUS-2626**
+  - Update [PDR migration](https://github.com/nasa/cumulus/blob/master/lambdas/data-migration2/src/pdrs.ts) to correctly find Executions by a Dynamo PDR's `execution` field
 
 ## [v9.3.0] 2021-07-26